refactor: extract case conversion to TypeMapper
Instead of having each builder be responsible for its own case
conversion, e.g. for type names, that responsibility is extracted and
moved into TypeMapper.

This is in preparation for adding the ability to change the type
conversion, or disable it entirely, via configuration.

By extracting out the responsibility for type name case conversion,
Generator can pass it in uniformly, rather than having that
responsibility diffused across many classes.

This does, however, muddy what TypeMapper's actual responsibility is,
however... so it might need to move from there or TypeMapper might need
to change to refelct that.