fix: implicit order dependency with custom types
Makes custom types work regardless of what order they are defined or
used by not building until all of the types have been registered.

Since we order types alphabetically, if a later type were to depend on
an earlier type (e.g. Table A uses Type Z), the type would be unknown at
the point when A was built and not available until after Z.

Instead of caching the built statements, we instead cache the builders
themselves during the build step and only after creating all builders
and registering their types do we build the nodes (including looking up
types in the registry).