feat: overhaul DBDataSource
This is basically a complete rewrite, but it does attempt to (mostly)
maintain API-compatibility for the moment. Once both projects using the
library are converted over to the new API, those deprecated functions
will be removed with prejudice.

There are some types that aren't quite compatible with the *declared*
types of the old API, but I think that in usage they will generally be
compatible, or at least very easily converted.