fix(generator): use case-insensitive collations
This changes the schema info queries to order by name in a
case-insensitive manner by lowercasing names in the order clause.

For better or worse, PostgreSQL uses the operating system's
implementation of collation, which leads to inconsistencies in ordering
across different operating systems.

This means that generated database types are not idempotent, leading to
issues in CI.

By lowercasing the names before ordering, we avoid the issue of
case-sensitive, operating system-dependent collations.

See-Also: https://dba.stackexchange.com/a/131471/134777