feat: transaction support
Adds rudimentary transaction support to DBDataSource

Previously, no support for transactions was available, leading to DBDS
only being able to be used for simplistic use-cases.

Adds a public `transaction` method to DBDataSource instances which will
start execute a given callback within the context of a transaction. That
transaction will apply to all DBDataSource instances with the same
database pool.

Unfortunately, the API is a little bit strange -- you can start the
transaction on any instance, and then use any other instance within it.
It could possibly be nice to pull out all of that handling into some
kind of transaction manager, but for now we're sticking with this simple
implementation, as it works as expected.

That said, be careful with loader cache priming -- any _new_ lookups
performed within the transaction will be cleared from the cache, but any
manual priming that is done (e.g. after running an insert or update)
will not be automatically cleared.