feat: add zod-based schema parsing capability
Previously, generated row types accounted for JSON/JSONB columns by
making them generic parameters to the row/insert types, but the actual
DataSource didn't really account for that at all.

Thus, when a caller had a generic type, they would have to do a bump of
hoop-jumping to actually verify the schema they are _claiming_ applies
to the column.

This change adds the ability to specify Zod schemas to be used to
validate JSON/JSONB columns. Those schemas are used at
INSERT/UPDATE-time, but *NOT* at SELECT time.

Parsing columns coming from the database may be something added in the
future, but for now it is much safer to avoid.