feat: generate columnTypes as const rather than Record
This changes the '$Types' object generation to generate `as const`
objects rather than `Record`.

Using `as const` allows the type system itself to know the types of
columns statically, which it previously could not infer because the
specific values were lost due to the generic `Record` use.

This is a necessary change to enable the JSON column schema parsing to
only allow callers to specify schemas for JSON/JSONB columns. It's also
just better, I think, but I was unaware of it at the time the feature
was originally written.